fix(cd): error msg
