+++ srcs/builtin/ft_cd/chdir.c
@@ -1,3 +1,4 @@
+#include <errno.h>
 #include <unistd.h>
 #include "minishell.h"
 #include "ms_builtin.h"
@@ -7,22 +8,24 @@ bool	is_absolute_path(const char *path)
 	return (path[0] == ABSOLUTE_PATH_HEAD);
 }
 
-static t_result	exec_chdir(const char *path, uint8_t *status)
+static t_result	exec_chdir(const char *path, int *tmp_err)
 {
+	errno = 0;
 	if (chdir(path) == CHDIR_ERROR)
 	{
-		*status = 1; // todo
+		*tmp_err = errno;
 		return (PROCESS_ERROR);
 	}
+	*tmp_err = errno;
 	return (SUCCESS);
 }
 
 // if chdir error, no need for auto perror.
 // "pwd + path" -> "path"
 t_result	cd_change_dir_to_valid_path(const char *absolute_path, \
-										uint8_t *status)
+										int *tmp_err)
 {
-	if (exec_chdir(absolute_path, status) == SUCCESS)
+	if (exec_chdir(absolute_path, tmp_err) == SUCCESS)
 		return (SUCCESS);
 	return (FAILURE);
 }
