--- srcs/error/error_msg.c
@@ -1,41 +1,0 @@
-#include "minishell.h"
-#include "ms_builtin.h"
-
-void	puterr_cmd_msg(const char *const cmd, const char *msg)
-{
-	ft_dprintf(STDERR_FILENO, \
-			"%s: %s: %s\n", \
-			SHELL_NAME, cmd, msg);
-
-}
-
-void	puterr_cmd_arg_msg(const char *cmd, \
-							const char *const arg, \
-							const char *msg)
-{
-	ft_dprintf(STDERR_FILENO, \
-			"%s: %s: %s: %s\n", \
-			SHELL_NAME, cmd, arg, msg);
-
-}
-
-void	puterr_msg_quoted_arg(const char *msg, const char *const arg)
-{
-	ft_dprintf(STDERR_FILENO, \
-				"%s: %s `%s'\n", \
-				SHELL_NAME, msg, arg);
-}
-
-void	puterr_arg_op_msg(const char *const arg, const char op)
-{
-	ft_dprintf(STDERR_FILENO, \
-				"%s: %s: %c%c: %s\n", \
-				SHELL_NAME, arg, CMD_OPTION_MARKER, op, ERROR_MSG_INVALID_OP);
-}
-
-void	puterr_heredoc_wanted_eof(const char *delimiter)
-{
-	ft_dprintf(STDERR_FILENO, \
-			"%s: %s: %s (wanted `%s')\n", \
-			SHELL_NAME, ERROR_TYPE_WARNING, ERROR_MSG_HEREDOC_EOF, delimiter);
-}
